fix(): remove operations.go file
